refactor: import buildCreateChildrenFn dynamically only when needed
